### 2023-03-06 #5223 @editorialbot
[PRE REVIEW]: PlatiPy: Processing Library and Analysis Toolkit for Medical Imaging in Python
**Submitting author:** @pchlap (<a href="http://orcid.org/0000-0002-6517-8745">Phillip Chlap</a>)
**Repository:** https://github.com/pyplati/platipy
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.6.0
