### 2023-03-15 f192858
Use Cosine annealing LR
