+++ pyproject.toml
@@ -36,7 +36,7 @@ Flask = { version = "^2.1.0", optional = true }
 Flask-RESTful = { version = ">=0.3.8", optional = true }
 Flask-SQLAlchemy = { version = "^2.4.4", optional = true }
 celery = { version = "^5.2.3", optional = true }
-redis = { version = "^3.5.3", optional = true }
+redis = { version = ">=3.5.3,<5.0.0", optional = true }
 psutil = { version = "^5.8.0", optional = true }
 gunicorn = { version = "^20.0.4", optional = true }
 Jinja2 = { version = "^3.1", optional = true }
