### 2023-05-17 1a43c14
Clean up code and prepare doc
