+++ docs/conf.py
@@ -72,6 +72,9 @@ shutil.copy("../examples/dvh_analysis.ipynb", "_examples/dvh_analysis.ipynb")
 shutil.copy("../examples/atlas_segmentation.ipynb", "_examples/atlas_segmentation.ipynb")
 shutil.copy("../examples/contour_comparison.ipynb", "_examples/contour_comparison.ipynb")
 shutil.copy("../examples/bronchus_segmentation.ipynb", "_examples/bronchus_segmentation.ipynb")
+shutil.copy(
+    "../examples/left_ventricle_17_segments.ipynb", "_examples/left_ventricle_17_segments.ipynb"
+)
 
 shutil.rmtree("site/assets", ignore_errors=True)
 os.makedirs("site", exist_ok=True)
