### 2023-05-31 #5374 @tomelse
Really nice work, sorry for taking so long to do the review. I only have one recommendation I think, which would be to add some more detail (and references) to the paper about existing applications that do similar tasks, particularly on the visualisation side. One example off the top of my head is MITK (https://www.mitk.org/wiki/The_Medical_Imaging_Interaction_Toolkit_(MITK)), which might be a good example to compare this to in your statement of need.
