### 2023-06-03 #5374 @agisga
I apologize for the delayed review. Overall, the `platipy` package appears to be in great shape. The documentation is sufficient, and the example Jupyter notebooks are well-designed and very helpful.

It is worth noting that `platipy` appears to be built on top of widely used lower-level imaging libraries, such as `pydicom`, `SimpleITK` and maybe others (at least in part). It may be beneficial to acknowledge these underlying packages and provide citations to the respective sources (e.g., I see citation of SimpleITK but not pydicom). However, discerning the exact extent to which `platipy` relies on `pydicom`, `SimpleITK`, or other libraries for its DICOM and image processing functionalities is somewhat challenging for me. It may be helpful to have a clearer understanding of these implementation details in the paper, but I leave it up to your judgement.

The JOSS review criteria (<https://joss.readthedocs.io/en/latest/review_criteria.html>) suggest including "A description of how this software compares to other commonly-used packages in this research area.". Thus, as also noted by the other reviewer, it would be beneficial to discuss alternative software solutions. For instance, in a paper on the Med-ImageTools Python package (<https://f1000research.com/articles/12-118>), the authors differentiate their solution from `platipy` and other alternatives. Although it is not a JOSS paper, it may be a useful reference related to this issue (note that I didn't read that paper in detail).

A few minor issues that I came across:
- Minor error when I run the last cell of `dvh_analysis.ipynb`: `AttributeError: 'NoneType' object has no attribute 'rowspan'` (Traceback points to `pandas/plotting/_matplotlib/tools.py:396`). But the visualization shows up anyway, so I didn't investigate further.
- I couldn't execute `generate_synthetic_head_neck_deformation.ipynb` due to: `SystemError: Animation functionality requires a matplotlib version lower than 3.4.0`. I'm sure it would be easy to just downgrade matplotlib, but I didn't try. Perhaps there is a better way to deal with this at installation time? Personally, I don't think that this is a significant issue.
