+++ paper.bib
@@ -28,6 +28,16 @@ volume = {31},
 year = {2018}
 }
 
+@misc{Zenodo:mason,
+  doi = {10.5281/ZENODO.7319790},
+  url = {https://zenodo.org/record/7319790},
+  author = {Mason,  Darcy and {Scaramallion} and {Mrbean-Bremen} and {Rhaxton} and Suever,  Jonathan and {Vanessasaurus} and Orfanos,  Dimitri Papadopoulos and Lemaitre,  Guillaume and Panchal,  Aditya and Rothberg,  Alex and Herrmann,  Markus D. and Massich,  Joan and Kerns,  James and {Korijn Van Golen} and Robitaille,  Thomas and Biggs,  Simon and {Moloney} and Bridge,  Chris and Shun-Shin,  Matthew and Conrad,  Blair and {Pawelzajdel} and Mattes,  Markus and Lyu,  YoungKi and Morency,  Félix C. and Cogan,  Tim and Sánchez,  Bernardo Pericacho and Meine,  Hans and Wortmann,  Joseph and Hahn,  Kevin S. and Wada,  Masahiro},
+  title = {pydicom/pydicom: pydicom 2.3.1},
+  publisher = {Zenodo},
+  year = {2022},
+  copyright = {Open Access}
+}
+
 @article{JMLR:v12:pedregosa11a,
   author  = {Fabian Pedregosa and Ga{{\"e}}l Varoquaux and Alexandre Gramfort and Vincent Michel and Bertrand Thirion and Olivier Grisel and Mathieu Blondel and Peter Prettenhofer and Ron Weiss and Vincent Dubourg and Jake Vanderplas and Alexandre Passos and David Cournapeau and Matthieu Brucher and Matthieu Perrot and {{\'E}}douard Duchesnay},
   title   = {Scikit-learn: Machine Learning in Python},
@@ -37,6 +47,62 @@ year = {2018}
   number  = {85},
   pages   = {2825-2830},
   url     = {http://jmlr.org/papers/v12/pedregosa11a.html}
+}
+
+@software{MITK_Team_MITK_2023,
+  author = {MITK Team},
+  license = {BSD-3-Clause},
+  month = may,
+  title = {{MITK}},
+  url = {https://github.com/MITK/MITK},
+  version = {v2023.04},
+  year = {2023}
+}
+
+@software{MIPAV,
+  author = {{National Institutes of Health}},
+  title = {MIPAV (Medical Image Processing, Analysis, and Visualization)},
+  url = {https://mipav.cit.nih.gov},
+  version = {11.0.7},
+  date = {2022-12-08},
+}
+
+@software{OnkoDICOM,
+  author = {{Maher, Ashley}},
+  title = {OnkoDICOM},
+  url = {https://onkodicom.com.au},
+  version = {0.05.02-beta},
+  date = {2022-11-02},
+}
+
+@article{Panchal2010,
+  doi = {10.1118/1.3468652},
+  url = {https://doi.org/10.1118/1.3468652},
+  year = {2010},
+  month = jun,
+  publisher = {Wiley},
+  volume = {37},
+  number = {6Part19},
+  pages = {3245--3245},
+  author = {A Panchal and R Keyes},
+  title = {{SU}-{GG}-T-260: Dicompyler: An Open Source Radiation Therapy Research Platform with a Plugin Architecture},
+  journal = {Medical Physics}
+}
+
+@article{Cardoso_MONAI_An_open-source_2022,
[…]
