+++ paper.md
@@ -30,11 +30,13 @@ bibliography: paper.bib
 
 # Summary
 
-PlatiPy offers a comprehensive suite of tools and utilities for conducting medical image analysis research utilising Python. These tools include functions for converting data between the clinical standard DICOM format and the research-friendly NIfTI format, capabilities for image registration and atlas-based segmentation, and efficient image visualsation tools to facilitate rapid development in research. Additionally, the library includes auto-segmentation models developed through various research projects enabling their streamlined deployment for use in future projects by the research community.
+PlatiPy offers a comprehensive suite of tools and utilities for conducting medical image analysis research utilising Python. These tools include functions for converting data between the clinical standard DICOM format and the research-friendly NIfTI format, capabilities for image registration and atlas-based segmentation, and efficient image visualisation tools to facilitate rapid development in research. Additionally, the library includes auto-segmentation models developed through various research projects enabling their streamlined deployment for use in future projects by the research community.
 
 # Statement of need
 
-Python has gained significant popularity in the field of medical image analysis research in recent years, due in part to its open-source nature and the support of a large community of third-party libraries. Libraries such as SimpleITK [@Lowekamp2013; @Yaniv2018] and scikit-learn [@JMLR:v12:pedregosa11a] offer a wide range of functionality for developing medical image analysis tools. However, researchers often face the challenge of writing code to prepare data, creating wrapper functions for common procedures, and visualising images with scalar, vector or structure overlays throughout the analysis pipeline. PlatiPy addresses these challenges by providing many of these functions, eliminating the need for researchers to reinvent the wheel.
+Python has gained significant popularity in the field of medical image analysis research in recent years, due in part to its open-source nature and the support of a large community of third-party libraries. Libraries such as SimpleITK [@Lowekamp2013; @Yaniv2018], scikit-learn [@JMLR:v12:pedregosa11a] and pydicom [@Zenodo:mason] offer a wide range of functionality for developing medical image analysis tools. However, researchers often face the challenge of writing code to prepare data, creating wrapper functions for common procedures, and visualising images with scalar, vector or structure overlays throughout the analysis pipeline. PlatiPy addresses these challenges by providing many of these functions, eliminating the need for researchers to reinvent the wheel.
+
+Other libraries and toolkits exist which provide comparable functionality to PlatiPy. For example the Medical Imaging Interaction Toolkit (MITK) [@MITK_Team_MITK_2023] provides a C++ library and a graphical to perform a wide range of medical image analysis tasks including reading DICOM data, performing image registration and producing visualisations. The MIPAV (Medical Image Processing, Analysis, and Visualization) application [@MIPAV], OnkoDICOM [@OnkoDICOM] and dicompyler [@Panchal2010] also provide a graphical user interface for medical image analysis and analysis of radiotherapy data. While these tool have proven useful in many research projects, they do not provide Python bindings which is often desired in modern medical image analysis research software projects. Medical Open Network for AI (MONAI) is an open-source framework for deep learning in healthcare imaging [@Cardoso_MONAI_An_open-source_2022]. This is built on top of the pytorch ecosystem and would typically be used to develop an entire project. In comparison, PlatiPy provides many utility type functions which can be integrated as little or as much as needed for a research project. The Medical Image Processing in Python library (medpy) does provide many of these utility type functions however doesn't provide functionality in the way of image visualisation or image registration available in PlatiPy [@medpy].
 
[…]
