+++ paper.bib
@@ -50,7 +50,7 @@ year = {2018}
 }
 
 @software{MITK_Team_MITK_2023,
-  author = {MITK Team},
+  author = {{MITK Team}},
   license = {BSD-3-Clause},
   month = may,
   title = {{MITK}},
