### 2023-06-08 #5374 @agisga
@samhforbes Yes, this addresses all of my open issues adequately.

@pchlap Minor grammar issue in the paper:

> [MITK] provides a C++ library and a graphical to perform

This should say "graphical user interface" instead of "graphical"?
