### 2023-06-09 #5374 @pchlap
> @pchlap Minor grammar issue in the paper:
>
> > [MITK] provides a C++ library and a graphical to perform
>
> This should say "graphical user interface" instead of "graphical"?

Thank you @agisga, I have corrected this grammar issue in the paper.
