### 2023-06-13 #5374 @pchlap
Thanks @samhforbes, I have addressed the author tasks in the post-review checklist:
- Deployed an updated version of PlatiPy: `0.6.1`
- This has been archived on Zenodo, DOI: `10.5281/zenodo.8032858`

Please let me know if there is anything else needed from me. Thanks!
