+++ paper.bib
@@ -162,6 +162,7 @@ keywords = {Automatic contouring,central tumors,contour variat,automatic contour
 pages = {94--101},
 publisher = {Elsevier Inc.},
 title = {{Medical Dosimetry Assessing tumor centrality in lung stereotactic ablative body radiotherapy ( SABR ): the effects of variations in bronchial tree delineation and potential for automated methods}},
+journal = {Medical Dosimetry},
 volume = {46},
 year = {2021}
 }
