### 2023-06-13 #5374 @pchlap
> Hi @pchlap I can see an issue with the bib: Ghandourh, W., Dowling, J., Chlap, P., Oar, A., Jacob, S., Batumalai, V., & Holloway, L. (2021). Medical Dosimetry Assessing tumor centrality in lung stereotactic ablative body radiotherapy ( SABR ): the effects of variations in bronchial tree delineation and potential for automated methods. 46, 94–101. https://doi.org/10.1016/j.meddos.2020.09.004
>
> * I think should be:
>
> Ghandourh, W., Dowling, J., Chlap, P., Oar, A., Jacob, S., Batumalai, V., & Holloway, L. (2021). Assessing tumor centrality in lung stereotactic ablative body radiotherapy ( SABR ): the effects of variations in bronchial tree delineation and potential for automated methods. _Medical Dosimetry, 46_, 94–101. https://doi.org/10.1016/j.meddos.2020.09.004
>
> Also in Cardoso et al https prefix comes twice

Thanks @samhforbes, I have corrected these two issues.
