### 2023-06-13 #5374 @samhforbes
Hi @pchlap unless I'm missing something, Ghandourh et al looks like it still has the Journal name in the article title.
