### 2023-06-13 bab8db3
Remove journal for article title
