+++ paper.bib
@@ -161,7 +161,7 @@ file = {:Users/60126181/OneDrive - UNSW/Mendeley/Ghandourh et al/Ghandourh et al
 keywords = {Automatic contouring,central tumors,contour variat,automatic contouring},
 pages = {94--101},
 publisher = {Elsevier Inc.},
-title = {{Medical Dosimetry Assessing tumor centrality in lung stereotactic ablative body radiotherapy ( SABR ): the effects of variations in bronchial tree delineation and potential for automated methods}},
+title = {{Assessing tumor centrality in lung stereotactic ablative body radiotherapy ( SABR ): the effects of variations in bronchial tree delineation and potential for automated methods}},
 journal = {Medical Dosimetry},
 volume = {46},
 year = {2021}
