### 2023-06-14 #5374 @pchlap
> Hi @pchlap this is all looking really good. I think you've done a great job with software and paper, and I'm going to hand over to the EiC now with a recommendation to accept!

Thank you @samhforbes, this has been a great experience. I have really enjoyed the JOSS review process, thanks a lot for coordinating! :)
